requirements.txt
anyio==4.11.0
apscheduler==3.11.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
google-api-core==2.26.0
google-api-python-client==2.88.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.14.0
httpcore==0.17.3
httplib2==0.31.0
httpx==0.24.1
idna==3.11
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.5
python-dotenv==1.0.1
python-telegram-bot==20.4
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.28
typing-extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20